Clean フォルダ分け後にHellow,Workldが動くことを確認
