+++ go/infrastructure/router.go
@@ -1,0 +1,27 @@
+package infrastructure
+
+import (
+	"net/http"
+
+	"github.com/labstack/echo"
+	"github.com/labstack/echo/middleware"
+)
+
+// Init Start webserver
+func Init() {
+	e := echo.New()
+
+	// ミドルウェア
+	e.Use(middleware.Logger())
+	e.Use(middleware.Recover())
+
+	// router
+	e.GET("/", mainPage())
+	e.Start(":3000") //ポート番号指定してね
+}
+
+func mainPage() echo.HandlerFunc {
+	return func(c echo.Context) error {
+		return c.JSON(http.StatusOK, "Hello World")
+	}
+}
